Bar - rename custom legend component
