+++ src/components/bar-chart/custom-bar-legend.js
